fixes in code for history and approvals. users can view their own movies/people even before approval

+++ app/controllers/api/v1/approvals_controller.rb
@@ -1,4 +1,4 @@
-class Api::V1::ApprovalsController < ApplicationController
+class Api::V1::ApprovalsController < Api::V1::BaseController
 
   respond_to :json
 
@@ -11,39 +11,45 @@ class Api::V1::ApprovalsController < ApplicationController
           original_id = params[:original_id]
           approved_id = params[:approved_id]
 
-          # get currently active item and duplicate it for backup
-          original_record = type.classify.constantize.where(id: original_id)
-          if original_record.count > 0
-            original_record = original_record.first
-            backup_status = backup_original(original_record)
+          if original_id == approved_id
+            original_record = type.classify.constantize.where(id: original_id).first
+            original_record.approved = true
+            original_record.save
+          else
+            # get currently active item and duplicate it for backup
+            original_record = type.classify.constantize.where(id: original_id)
+            if original_record.count > 0
+              original_record = original_record.first
+              backup_status = backup_original(original_record)
 
-            # get the approved record
-            approved_record = type.classify.constantize.where(id: approved_id)
-            if approved_record.count > 0
-              approved_record = approved_record.first
-              approved_values = approved_record.attributes
-              approved_values  = approved_values.except("id", "created_at", "updated_at")
-              approved_values["approved"] = mark
+              # get the approved record
+              approved_record = type.classify.constantize.where(id: approved_id)
+              if approved_record.count > 0
+                approved_record = approved_record.first
+                approved_values = approved_record.attributes
+                approved_values  = approved_values.except("id", "created_at", "updated_at")
+                approved_values["approved"] = mark
 
-              # update currently active item with new approved values
-              original_record.approved = true
-              if original_record.update_attributes(approved_values)
-                user = User.find approved_record.user_id
-                add_points_to_user(user)
-                add_badges_to_user(user)
-                approved_record.destroy
-                format.json { render json: original_record }
+                # update currently active item with new approved values
+                original_record.approved = true
+                if original_record.update_attributes(approved_values)
+                  user = User.find approved_record.user_id
+                  add_points_to_user(user)
+                  add_badges_to_user(user)
+                  approved_record.destroy
+                  format.json { render json: original_record }
+                else
+                  format.json { render :json => "Error approving content.", :status => :unprocessable_entity }
+                end
               else
-                format.json { render :json => "Error approving content.", :status => :unprocessable_entity }
+                if backup_status
+                  type.classify.constantize.last.destroy
+                end
+                format.json { render :json => "Error approving content not found.", :status => :unprocessable_entity }
               end
             else
-              if backup_status
-                type.classify.constantize.last.destroy
-              end
-              format.json { render :json => "Error approving content not found.", :status => :unprocessable_entity }
+              format.json { render :json => "Error original content not found.", :status => :unprocessable_entity }
             end
-          else
-            format.json { render :json => "Error original content not found.", :status => :unprocessable_entity }
           end
         else
           id = params[:approved_id]
